requirements.txt
flask
flask-socketio
python-socketio
mysql-connector-python
